Added support for verification with certificates
Now the tool can read x509 cert in pem format. -V/--verify-token-with now accept also cert file, the script will first treat the file as it's a public key, then, if reading the key return none, it try to load it as certificate and extracts the public key from it. If reading the file using the new x509 reading function will return none again it quits