Modified change_payload_complex, Corrected install.sh
Install.sh has an error in the modification of the PATH variable. Now it has been corrected. The build_keys method now return the string itself if it doesn't contain commas. In change_payload_complex the verification now does not check that keys is not None anymore, but it checks that keys is an instance of list containing more than one element