Added a feature to generate key from certs contained in x5c claims
As for public numbers, now we can generates keys from JWKs using the x5c with the method gen_public_key_from_x5c. In find_verifier_key_from_jwks, now the script first look for the claim 'x5c'. If its present the public key is generated with the using that new method. This creates anyway a new issue; find_verifier_key_from_jwks iterates a list of JWK while x5c is usually an array itself. So it seemed like it was necessary another feature to find the verifier cert in the x5c array. Anyway, according to the JWK standard, x5c verifier certificate should always be the first in the array. So the new method has only to check if the x5c is a string or an array, if it's a string it just uses that string, if it's an array, uses the first object contained in the array.
Now in x5u_basic and x5u_via_header_injection, the tool will also add public numbers of the public key