Added checks of x5c type in x5u methods
According to last commit, now also x5u_basic and x5u_header_injection methods, looks for the x5c type. If is list, our crafted x5c is inserted as the first element in the list, else, if it's not a list, our crafted x5c is set as x5c value (string).