Modified jwk claim generation and embedding according to instructions in the improvements queue
Now jwk claim can be generated for all supported jwa. Insertiing a jwk claim in the token header it's now done via a proper method, that checks if it already contain a keys set (looking for the claim 'keys'). If exists, instead of overriding it, the method insert the new jwk as the 1st element of the keys set.