Specified an output filename in jku/x5u functions, for download jwks files
Sine now jwks files was downloaded with theri original filename, so the tool was loading the first .json file it found in the current directory. This potentially generates two issues: First, if current dir already contains a .json file, the tool could load the wrong one (sine it list in alphabetical order), then, if the file on the server does not end with .json, the tool may miss to load it, returning an error. Now a custom filename 'jwtxpl_jwks.json' has been set and passed to wget using the '-O/--output' option. This should solve the issue