Added support for key reading in x5u injections
Now with x5u injection can be specified a key to sign the certificate with. This implementation only consist in an if/else statement to check if self.path_to_key has a value when looking for x5u argument in self.validation. Now the x5u_command variable is defined differently if self.path_to_key is None or not. If self.path_to_key has None value (actually is 'if not self.path_to_key' and not 'if self.path_to_key is None', but self.path_to_key will never be an empty string) the command requiring a new cert, generates also a new key, else, the file it's pointing to, will be used to sign the cert.