Corrected error in jwks file generation in x5u_basic method
The x5u_basic method was generating a file called {CWD}jwks.json in the current directory, instead of generating it under the repo crafted directory and naming it only jwks.json. That was due to a missing sting formatting, since the path to the crafted/ directory is stored in the CWD variable.