Added handling of ambiguous data types in payload changes
Until now, numeric strings in payload changes was always converted to integers. Now, preceding a numeric string with a '%' sign, will return a the numeric string itself (stripped of the '%' sign). This is due, so the user is able to preserve the original data type for a specific value. Wiki updates are following