Added key dumping as part of validation
the option and methods for key dumpign have been provided. Now they are handled in validation, returning proper warnings. For example, key dumping is not compatible with key reading, since the tool would read the key from a file and then override the file with the same content (useless resource consuming). The tool now also advise users that does not provides a key, that generating keys for each attack is useless and resource comsuming too, so they should use -D/--dump-key to generate a new keyfile and then use it in next attacks with -k/--key