Handled exception in new auto_try feature
Two exception of the new auto_try feature (see previous commit), has been handled. This require to import the socket library. If exception to handle will get more, maybe a singl 'except Exception' statement should be used, even if it's going to be not pep8 compliant (there's so many things here that are not).