Removed the spam of the warning related to complex_payload
Since the deprecation status of complex_payload method, a warning about it has been placed in the code, spawning when users call that method. It was placed in a for loop so the warning spawned as many times as complex_payload was called. It now has been moved outside the for loop