Defined conditional printing function
Preceding the new feature token_only --token-only, that it's going to ignore warnings and other infos, printing out only the generated JWT, a method called ifprint has been defined. It takes two parameter: condition and string. If condition is true, the function prints out string. So we can pass the new option token_only as condition to determine what to print. In our secific case the condition has to be passed as negative (not <condition>)