Added a method to extract the alg from a JWT header
In order to avoid the user having to spiecify the algortihm during a token verification, a method to extract the alg from the original base64 encoded JWT header. This is due since during validation we can't access the token dictionary (the dictionarize_token method is called after validation is done and we can't do it before, or we can run through errors since token has not been validated yet). Using this function cuold sound a bit of repetitive, the only alternative is to declare the variables containing the splitted and decoded token information in the init method, before validation run, and than initialize them during validation. Anyway this will require an extra step, casting those infomration to dict type, that's not included in the current process. Last, a new important feature is going to be implemented with next releases, that will make use of this new method too