Updated version for today merge, stopped verification to require the user to specify an algorithm
Since now, verifying operations, required the user to specify an algorithm. This was not the way since, in order to perform a reliable verification, every change in the token should be ignored. After some test, this is going to be extended to the find_verifier_key_from_jwks_and_quit method too