Implemented methods to use for public key computation
New release it's going to introduce a feature for compute public keys given 2 or more JWTs as input. So here, I'm introducing three methods that this feature will require: 1) Message hashing: A method to hash a message using hashlib, based on the 'alg' param of the token. 2) Messagge padding: A method to apply PKCS1_v15_EMSA_ENCODING to an hashed message. 3) OID selection: A method returning the hash OID. In order to recover the public key (that's going to be used in an algorithm confusion attack), we need to apply math on pairs of messsage and signature (let's refer to them as m:s). Firsts two parts of JWT are the message, the last one is the signature. Anyway, using 'header.payload' as message, will result in an invalid computation. This is caused by real signing methods, that apply hashing and padding methods to the message before signing it. So in order to have the real m of the m:s pair, we need to hash and apply padding (in this order) to the firsts two parts of the JWT.